move mouse events to RaycasterGroupDirective
